refactor: rename InMemoryCache and move to top-level package as it's generic, not exchange-specific
