+++ src/main/java/pl/cleankod/cache/CacheInMemoryClient.java
@@ -1,15 +1,17 @@
-package pl.cleankod.exchange.core.cache;
+package pl.cleankod.cache;
+
+import pl.cleankod.cache.inmemory.CacheInMemory;
 
 import java.util.Map;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Function;
 
-public class InMemoryCachingClient<K, V> {
+public class CacheInMemoryClient<K, V> {
     private final Function<K, V> fetchFunction;
     private final Map<K, V> cache;
     private final ReentrantLock lock;
 
-    public InMemoryCachingClient(Function<K, V> fetchFunction, InMemoryCache<K, V> cache) {
+    public CacheInMemoryClient(Function<K, V> fetchFunction, CacheInMemory<K, V> cache) {
         this.fetchFunction = fetchFunction;
         this.cache = cache;
         this.lock = new ReentrantLock();
