+++ src/main/java/pl/cleankod/cache/inmemory/CacheInMemoryLru.java
@@ -1,17 +1,17 @@
-package pl.cleankod.exchange.core.cache;
+package pl.cleankod.cache.inmemory;
 
 import java.util.Map;
 
-public class InMemoryLruCache<K, V> extends InMemoryCache<K, V> {
+public class CacheInMemoryLru<K, V> extends CacheInMemory<K, V> {
     private final int maxSize;
 
-    public InMemoryLruCache(int maxSize) {
+    public CacheInMemoryLru(int maxSize) {
         super(maxSize, 0.75f, true);
         this.maxSize = maxSize;
     }
 
     @Override
     protected boolean removeEldestEntry(Map.Entry<K, V> eldest) {
-        return size() > maxSize;
+        return this.size() > this.maxSize;
     }
 }
